# src/main/resources/api/test-api-swagger.yaml
swagger: "2.0"
info:
  description: "Esto es un test de Cliente Api. 
  [GitHub](https://github.com/Santiago8995/spring-apicliente-jpa)."
  version: "1.0.0"
  title: "Swagger Cliente Api 1.0"
  termsOfService: "https://github.com/Santiago8995/spring-apicliente-jpa"
  contact:
    email: "santiago.8.9.95@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "testapi-jpa.rj.r.appspot.com"
basePath: "/api/v1/clientes"
tags:
  - name: "Clientes"
    description: "Operaciones sobre Clientes"
schemes:
  - "http"
  - "https"
paths:
  /crearcliente:
    post:
      tags:
        - "cliente"
      summary: "Añade un nuevo cliente"
      description: "Crear un nuevo cliente en la base de datos."
      operationId: "crearcliente"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Datos necesarios para crear un nuevo cliente."
          required: true
          schema:
            properties:
              nombre:
                type: "string"
                example: Santiago
              apellido:
                type: "string"
                example: Pineiro
              fechaNacimiento:
                type: string
                example: 08-09-1995
      responses:
        "201":
          description: "CREATED (El cliente se creo correctamente)"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int32"
                example: 1
              nombre:
                type: "string"
                example: "Santiago"
              apellido:
                type: "string"
                example: "Pineiro"
              fechaNacimiento:
                type: "string"
                example:  08-09-1995

        "400":
          description: "BAD REQUEST - Error en el formato de alguno de los datos enviados."
  /listarclientes:
    get:
      tags:
        - "cliente"
      summary: "Encuentra una lista de clientes"
      description: "Devuelve una lista de clientes, con sus edades y su tiempo estimado de muerte(Segun el promedio global de vida de un ser humano)."
      operationId: "listarclientes"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Cliente"
        "404":
          description: "NOT FOUND"
  /kpicliente:
    get:
      tags:
        - "cliente"
      summary: "Genera un KPI de edades"
      description: "Crea un dato de Promedio y otro de Desviacion Estandar en base a las edades de todos los clientes guardados."
      operationId: "kpiclientes"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/KpiClientes"
        "400":
          description: "Invalid tag value"
definitions:
  Cliente:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      nombre:
        type: "string"
      apellido:
        type: "string"
      fechaNacimiento:
        type: "string"
      edad:
        type: "integer"
        format: "int64"
      vidaEstimada:
        type: "integer"
        format: "int64"
  KpiClientes:
    type: "object"
    properties:
      kpiCliente:
        type: "number"
      desvEstandar:
        type: "number"